Compute Pearson or Spearman correlation coefficients

Evaluate the relationship between a numeric outcome and a numeric predictor, by computing the Pearson or Spearman correlation coefficients. Output a tibble result with with one row per predictor, and four columns: name, score, predictor, and outcome.

Usage

get_scores_cor(score_obj, data, outcome)

Arguments

score_obj

A score object. See score_cor() for details.

data

A data frame or tibble containing the outcome and predictor variables.

outcome

A character string specifying the name of the outcome variable.

Value

A tibble of result with one row per predictor, and four columns:

name: the name of scoring metric.

score: the score for the predictor-outcome pair.

predictor: the name of the predictor.

outcome: the name of the outcome.

Examples

data(ames, package = "modeldata")
ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
# Return score as pearson correlation
score_obj <- score_cor()
score_res <- get_scores_cor(
  score_obj,
  data = ames_subset,
  outcome = "Sale_Price"
)
score_res
#> # A tibble: 5 × 4
#>   name     score outcome    predictor   
#>   <chr>    <dbl> <chr>      <chr>       
#> 1 pearson NA     Sale_Price MS_SubClass 
#> 2 pearson NA     Sale_Price MS_Zoning   
#> 3 pearson  0.202 Sale_Price Lot_Frontage
#> 4 pearson  0.267 Sale_Price Lot_Area    
#> 5 pearson NA     Sale_Price Street      
# Return score as spearman correlation
score_obj <- score_cor(score_type = "spearman")
score_res <- get_scores_cor(
  score_obj,
  data = ames_subset,
  outcome = "Sale_Price"
)
score_res
#> # A tibble: 5 × 4
#>   name      score outcome    predictor   
#>   <chr>     <dbl> <chr>      <chr>       
#> 1 spearman NA     Sale_Price MS_SubClass 
#> 2 spearman NA     Sale_Price MS_Zoning   
#> 3 spearman  0.228 Sale_Price Lot_Frontage
#> 4 spearman  0.429 Sale_Price Lot_Area    
#> 5 spearman NA     Sale_Price Street