Scoring via correlation coefficient

These objects are used when:

The predictors are numeric and the outcome is numeric.

In this case, a correlation coefficient (via stats::cor()) is computed with the proper variable roles. Values closer to 1 or -1 (i.e., abs(cor_pearson) closer to 1) are associated with more important predictors.

score_cor_pearson and score_cor_spearman are objects that define the technique. To apply the filter on data, you would use the fit() method:

  fit(score_cor_pearson, formula, data)

See the Examples section below.

Usage

score_cor_pearson

score_cor_spearman

Format

An object of class filtro::class_score_cor (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_cor (inherits from filtro::class_score, S7_object) of length 1.

Arguments

object

A score class object based on class_score_cor.

formula

A standard R formula with a single outcome on the right-hand side and one or more predictors (or .) on the left-hand side. The data are processed via stats::model.frame().

data

A data frame containing the relevant columns defined by the formula.

case_weights

A quantitative vector of case weights that is the same length as the number of rows in data. The default of NULL indicates that there are no case weights.

...

Further arguments passed to or from other methods.

Details

The function will determine which columns are predictors and outcomes and compute correlations; no user intervention is required.

Missing values are removed for each predictor/outcome combination being scored. In cases where cor() fail, the scoring proceeds silently, and a missing value is given for the score.

Examples

library(dplyr)

ames_subset <- modeldata::ames |>
  select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )

ames_cor_pearson_res <-
  score_cor_pearson |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_cor_pearson_res@results
#> # A tibble: 5 × 4
#>   name         score outcome    predictor   
#>   <chr>        <dbl> <chr>      <chr>       
#> 1 cor_pearson NA     Sale_Price MS_SubClass 
#> 2 cor_pearson NA     Sale_Price MS_Zoning   
#> 3 cor_pearson  0.202 Sale_Price Lot_Frontage
#> 4 cor_pearson  0.267 Sale_Price Lot_Area    
#> 5 cor_pearson NA     Sale_Price Street      

ames_cor_spearman_res <-
  score_cor_spearman |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_cor_spearman_res@results
#> # A tibble: 5 × 4
#>   name          score outcome    predictor   
#>   <chr>         <dbl> <chr>      <chr>       
#> 1 cor_spearman NA     Sale_Price MS_SubClass 
#> 2 cor_spearman NA     Sale_Price MS_Zoning   
#> 3 cor_spearman  0.228 Sale_Price Lot_Frontage
#> 4 cor_spearman  0.429 Sale_Price Lot_Area    
#> 5 cor_spearman NA     Sale_Price Street