Scoring via the chi-squared test or Fisher's exact test

These objects are used when:

The predictors are factors and the outcome is a factor.

In this case, a contingency table (via table()) is created with the proper variable roles, and the cross tabulation p-value is computed using either the chi-squared test (via stats::chisq.test()) or Fisher's exact test (via stats::fisher.test()). The p-value that is returned is transformed to be -log10(p_value) so that larger values are associated with more important predictors.

score_xtab_pval_chisq and score_xtab_pval_fisher are objects that define the technique. To apply the filter on data, you would use the fit() method:

  fit(score_xtab_pval_chisq, formula, data)

See the Examples section below.

Usage

score_xtab_pval_chisq

score_xtab_pval_fisher

Format

An object of class filtro::class_score_xtab (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_xtab (inherits from filtro::class_score, S7_object) of length 1.

Arguments

object

A score class object based on class_score_xtab.

formula

A standard R formula with a single outcome on the right-hand side and one or more predictors (or .) on the left-hand side. The data are processed via stats::model.frame().

data

A data frame containing the relevant columns defined by the formula.

...

Further arguments passed to or from other methods.

Details

The function will determine which columns are predictors and outcomes in the contingency table; no user intervention is required.

Missing values are removed for each predictor/outcome combination being scored. In cases where stats::chisq.test() or stats::fisher.test() fail, the scoring proceeds silently, and a missing value is given for the score.

Examples

# Binary factor example

library(titanic)
library(dplyr)

titanic_subset <- titanic_train |>
  mutate(across(c(Survived, Pclass, Sex, Embarked), as.factor)) |>
  select(Survived, Pclass, Sex, Age, Fare, Embarked)

titanic_xtab_pval_chisq_res <- score_xtab_pval_chisq |>
  fit(Survived ~ ., data = titanic_subset)
titanic_xtab_pval_chisq_res@results
#> # A tibble: 5 × 4
#>   name            score outcome  predictor
#>   <chr>           <dbl> <chr>    <chr>    
#> 1 xtab_pval_chisq 22.3  Survived Pclass   
#> 2 xtab_pval_chisq 57.9  Survived Sex      
#> 3 xtab_pval_chisq NA    Survived Age      
#> 4 xtab_pval_chisq NA    Survived Fare     
#> 5 xtab_pval_chisq  5.79 Survived Embarked 

titanic_xtab_pval_fisher_res <- score_xtab_pval_fisher |>
  fit(Survived ~ ., data = titanic_subset)
titanic_xtab_pval_fisher_res@results
#> # A tibble: 5 × 4
#>   name             score outcome  predictor
#>   <chr>            <dbl> <chr>    <chr>    
#> 1 xtab_pval_fisher 22.5  Survived Pclass   
#> 2 xtab_pval_fisher 59.2  Survived Sex      
#> 3 xtab_pval_fisher NA    Survived Age      
#> 4 xtab_pval_fisher NA    Survived Fare     
#> 5 xtab_pval_fisher  5.99 Survived Embarked 
# TODO Add multiclass example