Scoring via random forests

These objects are used when either:

The predictors are numeric and the outcome is a factor/category, or

The predictors are factors and the outcome is numeric.

In either case, a random forest, conditional random forest, or oblique random forest (via ranger::ranger(), partykit::cforest(), or aorsf::orsf()) is created with the proper variable roles, and the feature importance scores are computed. Larger values are associated with more important predictors.

score_imp_rf, score_imp_rf_conditional and score_imp_rf_oblique are objects that define the technique. To apply the filter on data, you would use the fit() method:

  fit(score_imp_rf, formula, data)

See the Examples section below.

Usage

score_imp_rf

score_imp_rf_conditional

score_imp_rf_oblique

Format

An object of class filtro::class_score_imp_rf (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_imp_rf (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_imp_rf (inherits from filtro::class_score, S7_object) of length 1.

Arguments

object

A score class object based on class_score_imp_rf.

formula

A standard R formula with a single outcome on the right-hand side and one or more predictors (or .) on the left-hand side. The data are processed via stats::model.frame().

data

A data frame containing the relevant columns defined by the formula.

...

Further arguments passed to or from other methods.

Details

The function will determine which columns are predictors and outcomes in the random forest; no user intervention is required.

Missing values are removed for each predictor/outcome combination being scored. When a predictor's importance score is 0, partykit::cforest() may omit its name from the results. In cases like these, a score of 0 is assigned to the missing predictors.

Examples

library(dplyr)

# Random forests for classification task

cells_subset <- modeldata::cells |>
  dplyr::select(
    class,
    angle_ch_1,
    area_ch_1,
    avg_inten_ch_1,
    avg_inten_ch_2,
    avg_inten_ch_3
  )

# Random forest
set.seed(42)
cells_imp_rf_res <- score_imp_rf |>
  fit(class ~ ., data = cells_subset)
cells_imp_rf_res@results
#> # A tibble: 5 × 4
#>   name       score outcome predictor     
#>   <chr>      <dbl> <chr>   <chr>         
#> 1 imp_rf 0.0000947 class   angle_ch_1    
#> 2 imp_rf 0.0318    class   area_ch_1     
#> 3 imp_rf 0.0672    class   avg_inten_ch_1
#> 4 imp_rf 0.0838    class   avg_inten_ch_2
#> 5 imp_rf 0.0104    class   avg_inten_ch_3

# Conditional random forest
cells_imp_rf_conditional_res <- score_imp_rf_conditional |>
  fit(class ~ ., data = cells_subset)
cells_imp_rf_conditional_res@results
#> # A tibble: 5 × 4
#>   name                 score outcome predictor     
#>   <chr>                <dbl> <chr>   <chr>         
#> 1 imp_rf_conditional 0.00332 class   angle_ch_1    
#> 2 imp_rf_conditional 0.0519  class   area_ch_1     
#> 3 imp_rf_conditional 0.0854  class   avg_inten_ch_1
#> 4 imp_rf_conditional 0.216   class   avg_inten_ch_2
#> 5 imp_rf_conditional 0.0169  class   avg_inten_ch_3

# Oblique random forest
cells_imp_rf_oblique_res <- score_imp_rf_oblique |>
  fit(class ~ ., data = cells_subset)
cells_imp_rf_oblique_res@results
#> # A tibble: 5 × 4
#>   name              score outcome predictor     
#>   <chr>             <dbl> <chr>   <chr>         
#> 1 imp_rf_oblique -0.00194 class   angle_ch_1    
#> 2 imp_rf_oblique  0.0548  class   area_ch_1     
#> 3 imp_rf_oblique  0.125   class   avg_inten_ch_1
#> 4 imp_rf_oblique  0.136   class   avg_inten_ch_2
#> 5 imp_rf_oblique  0.0147  class   avg_inten_ch_3

# ----------------------------------------------------------------------------

# Random forests for regression task

ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
ames_subset <- ames_subset |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))

regression_task <- score_imp_rf
regression_task@mode <- "regression"

set.seed(42)
ames_imp_rf_regression_task_res <-
  regression_task |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_imp_rf_regression_task_res@results
#> # A tibble: 5 × 4
#>   name       score outcome    predictor   
#>   <chr>      <dbl> <chr>      <chr>       
#> 1 imp_rf 0.0149    Sale_Price MS_SubClass 
#> 2 imp_rf 0.0112    Sale_Price MS_Zoning   
#> 3 imp_rf 0.00693   Sale_Price Lot_Frontage
#> 4 imp_rf 0.0143    Sale_Price Lot_Area    
#> 5 imp_rf 0.0000336 Sale_Price Street      
# TODO Add example of how to change trees, mtry, min_n, seed