Scoring via entropy-based filters

These objects are used when either:

The predictors are numeric and the outcome is a factor/category, or

The predictors are factors and the outcome is numeric.

In either case, an entropy-based filter (via FSelectorRcpp::information_gain()) is applied with the proper variable roles. Depending on the chosen method, information gain, gain ratio, or symmetrical uncertainty is computed. Larger values are associated with more important predictors.

score_info_gain, score_gain_ratio and score_sym_uncert are objects that define the technique. To apply the filter on data, you would use the fit() method:

fit(score_info_gain, formula, data)

See the Examples section below.

Usage

score_info_gain

score_gain_ratio

score_sym_uncert

Format

An object of class filtro::class_score_info_gain (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_info_gain (inherits from filtro::class_score, S7_object) of length 1.

An object of class filtro::class_score_info_gain (inherits from filtro::class_score, S7_object) of length 1.

Arguments

object

A score class object based on score_info_gain.

formula

A standard R formula with a single outcome on the right-hand side and one or more predictors (or .) on the left-hand side. The data are processed via stats::model.frame().

data

A data frame containing the relevant columns defined by the formula.

...

Further arguments passed to or from other methods.

Details

The function will determine which columns are predictors and outcomes in the filter; no user intervention is required.

Missing values are removed for each predictor/outcome combination being scored. When a predictor's importance score is 0, the gain ratio method may omit its name from the results. In cases like these, a score of 0 is assigned to the missing predictors.

Examples

rlang::is_installed("modeldata")
#> [1] TRUE

library(dplyr)

# Entropy-based filter for classification tasks

cells_subset <- modeldata::cells |>
  dplyr::select(
    class,
    angle_ch_1,
    area_ch_1,
    avg_inten_ch_1,
    avg_inten_ch_2,
    avg_inten_ch_3
  )

# Information gain
cells_info_gain_res <- score_info_gain |>
  fit(class ~ ., data = cells_subset)
cells_info_gain_res@results
#> # A tibble: 5 × 4
#>   name      score outcome predictor     
#>   <chr>     <dbl> <chr>   <chr>         
#> 1 infogain 0      class   angle_ch_1    
#> 2 infogain 0.0144 class   area_ch_1     
#> 3 infogain 0.109  class   avg_inten_ch_1
#> 4 infogain 0.137  class   avg_inten_ch_2
#> 5 infogain 0      class   avg_inten_ch_3

# Gain ratio
cells_gain_ratio_res <- score_gain_ratio |>
  fit(class ~ ., data = cells_subset)
cells_gain_ratio_res@results
#> # A tibble: 5 × 4
#>   name       score outcome predictor     
#>   <chr>      <dbl> <chr>   <chr>         
#> 1 gainratio 0      class   angle_ch_1    
#> 2 gainratio 0.0266 class   area_ch_1     
#> 3 gainratio 0.0828 class   avg_inten_ch_1
#> 4 gainratio 0.106  class   avg_inten_ch_2
#> 5 gainratio 0      class   avg_inten_ch_3

# Symmetrical uncertainty
cells_sym_uncert_res <- score_sym_uncert |>
  fit(class ~ ., data = cells_subset)
cells_sym_uncert_res@results
#> # A tibble: 5 × 4
#>   name       score outcome predictor     
#>   <chr>      <dbl> <chr>   <chr>         
#> 1 symuncert 0      class   angle_ch_1    
#> 2 symuncert 0.0242 class   area_ch_1     
#> 3 symuncert 0.111  class   avg_inten_ch_1
#> 4 symuncert 0.141  class   avg_inten_ch_2
#> 5 symuncert 0      class   avg_inten_ch_3

# ----------------------------------------------------------------------------

# Entropy-based filter for regression tasks

ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
ames_subset <- ames_subset |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))

regression_task <- score_info_gain
regression_task@mode <- "regression"

ames_info_gain_regression_task_res <-
  regression_task |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_info_gain_regression_task_res@results
#> # A tibble: 5 × 4
#>   name       score outcome    predictor   
#>   <chr>      <dbl> <chr>      <chr>       
#> 1 infogain 0.266   Sale_Price MS_SubClass 
#> 2 infogain 0.113   Sale_Price MS_Zoning   
#> 3 infogain 0.146   Sale_Price Lot_Frontage
#> 4 infogain 0.140   Sale_Price Lot_Area    
#> 5 infogain 0.00365 Sale_Price Street