Scoring via area under the Receiver Operating Characteristic curve (ROC AUC)

These objects are used when either:

The predictors are numeric and the outcome is a factor/category, or

The predictors are factors and the outcome is numeric.

In either case, a ROC curve (via pROC::roc()) is created with the proper variable roles, and the area under the ROC curve is computed (via pROC::auc()). Values higher than 0.5 (i.e., max(roc_auc, 1 - roc_auc) > 0.5) are associated with more important predictors.

score_roc_auc is object that define the technique. To apply the filter on data, you would use the fit() method:

  fit(score_roc_auc, formula, data)

See the Examples section below.

Usage

score_roc_auc

Format

An object of class filtro::class_score_roc_auc (inherits from filtro::class_score, S7_object) of length 1.

Arguments

object

A score class object based on class_score_roc_auc.

formula

A standard R formula with a single outcome on the right-hand side and one or more predictors (or .) on the left-hand side. The data are processed via stats::model.frame().

data

A data frame containing the relevant columns defined by the formula.

...

Further arguments passed to or from other methods.

Details

The function will determine which columns are predictors and outcomes for the ROC analysis; no user intervention is required.

Missing values are removed for each predictor/outcome combination being scored. In cases where pROC::roc() fail, the scoring proceeds silently, and a missing value is given for the score.

Examples

library(dplyr)

# ROC AUC where the numeric predictors are the predictors and
# `class` is the class outcome/response

cells_subset <- modeldata::cells |>
  dplyr::select(
    class,
    angle_ch_1,
    area_ch_1,
    avg_inten_ch_1,
    avg_inten_ch_2,
    avg_inten_ch_3
  )

cells_roc_auc_res <- score_roc_auc |>
  fit(class ~ ., data = cells_subset)
cells_roc_auc_res@results
#> # A tibble: 5 × 4
#>   name    score outcome predictor     
#>   <chr>   <dbl> <chr>   <chr>         
#> 1 roc_auc 0.502 class   angle_ch_1    
#> 2 roc_auc 0.591 class   area_ch_1     
#> 3 roc_auc 0.760 class   avg_inten_ch_1
#> 4 roc_auc 0.777 class   avg_inten_ch_2
#> 5 roc_auc 0.513 class   avg_inten_ch_3

# ----------------------------------------------------------------------------

# ROC AUC where `Sale_Price` is the numeric predictor and the class predictors
# are the outcomes/responses

ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
ames_subset <- ames_subset |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))

ames_roc_auc_res <- score_roc_auc |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_roc_auc_res@results
#> # A tibble: 5 × 4
#>   name     score outcome    predictor   
#>   <chr>    <dbl> <chr>      <chr>       
#> 1 roc_auc  0.742 Sale_Price MS_SubClass 
#> 2 roc_auc  0.853 Sale_Price MS_Zoning   
#> 3 roc_auc NA     Sale_Price Lot_Frontage
#> 4 roc_auc NA     Sale_Price Lot_Area    
#> 5 roc_auc  0.807 Sale_Price Street      
# TODO Add multiclass example